Fix tf2 gamedata not being used
When this was split out from the legacy obv gamedata in 5678c55 the new
block accidentally used the engine name for tf2 instead of the gamedir
in a gamedir match. Change it to use the same #supported rules as the
others to fix the gamedata loading and avoid future confusion.